refactor: Move code_graph tool logic into MCP server package
Relocate all modules from src/agentic_platform/tool/code_graph/ into
src/agentic_platform/mcp_server/code_graph_mcp_server/tool/ to colocate
the tool logic with the MCP server that consumes it.

Updated all import paths in server.py, parser.py, and branch_extractor.py.
Relative imports within the tool package (cache.py, query_tools.py) required
no changes.
